- bug in guessing FP version for a corresponding AIR version
git-svn-id: http://flashdevelop.googlecode.com/svn/trunk@2222 2354be2f-a62e-0410-afb7-89360a396392

+++ FD4/External/Plugins/ProjectManager/Projects/AS3/AS3Project.cs
@@ -268,7 +268,7 @@ namespace ProjectManager.Projects.AS3
 
         static public void GuessFlashPlayerForAIR(ref int majorVersion, ref int minorVersion)
         {
-            double v = majorVersion + minorVersion / 10;
+            double v = majorVersion + (double)minorVersion / 10;
             if (v < 2) { majorVersion = 9; minorVersion = 0; }
             else if (v < 2.5) { majorVersion = 10; minorVersion = 0; }
             else if (v < 2.6) { majorVersion = 10; minorVersion = 1; }
